requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
Pillow==10.1.0
PyMuPDF==1.23.8
python-dotenv==1.0.0
aiofiles==23.2.1
redis==5.0.1
pydantic==2.5.0
pydantic-settings==2.1.0
APScheduler==3.10.4
structlog==23.2.0
psutil==5.9.6
httpx==0.25.2
